requirements.txt
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
Click==7.0
cmake==3.13.3
dj-database-url==0.5.0
Django==2.1.7
django-crispy-forms==1.6.0
django-debug-toolbar==1.10.1
django-markdown-deux==1.0.5
django-markdown2==0.3.1
django-model-utils==3.1.2
django-notifications-hq==1.5.0
django-pagedown==0.1.1
djangorestframework==3.9.2
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Faker==1.0.2
future==0.17.1
gunicorn==19.9.0
idna==2.6
jsonfield==2.0.2
Markdown==2.6.11
markdown2==2.3.5
mathparse==0.1.2
nltk==3.4
numpy==1.16.2
oauthlib==2.1.0
opencv-python==4.0.0.21
Pillow==5.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.6.1
pycparser==2.19
pymongo==3.7.2
python-dateutil==2.6.1
python-twitter==3.5
pytz==2018.4
PyYAML==3.13
requests==2.18.4
requests-oauthlib==1.1.0
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.2.15
sqlparse==0.2.4
text-unidecode==1.2
urllib3==1.22
whitenoise==4.1
